+++ src/store/index.js
@@ -292,6 +292,13 @@ export default new Vuex.Store({
         state.helpAlert[key] = true
       }
     },
+    migratedAddLastViewedDate(state) {
+      const newVideoSettings = { ...state.videoSettings }
+      for (const videoSettings of Object.values(state.videoSettings)) {
+        videoSettings.lastViewed = new Date().toISOString()
+      }
+      state.videoSettings = newVideoSettings
+    },
   },
   plugins: [
     VuexWebExtensions({